backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# DB
python-dotenv==1.0.0

# ML (PIN THIS)
tensorflow==2.15.0

# Testing
pytest==7.4.3
httpx==0.25.1
